feat: add errors and style and labels
