+++ src/App.tsx
@@ -1,9 +1,11 @@
+import { FormRoom } from "./components/Register/formRoom";
 import { RoutesMain } from "./routes/routes";
 
 const App = () => {
   return (
     <>
       <RoutesMain />
+      <FormRoom></FormRoom>
     </>
   );
 };
