--- src/components/Register/index.tsx
@@ -1,3 +1,0 @@
-export const index = () => {
-  return <span>Apagar arquvio index.tsx</span>;
-};
